Extract Network and Interface from utils
