+++ tests/integration_cases/test_raft_paper.rs
@@ -27,6 +27,7 @@
 
 use protobuf::RepeatedField;
 use raft::eraftpb::*;
+use raft::harness::*;
 use raft::storage::MemStorage;
 use raft::*;
 use test_util::*;
