+++ tests/integration_cases/test_raw_node.rs
@@ -27,6 +27,7 @@
 
 use protobuf::{self, ProtobufEnum};
 use raft::eraftpb::*;
+use raft::harness::*;
 use raft::storage::MemStorage;
 use raft::*;
 use test_util::*;
